Prepare drivetrain commons for hot-loading config
tl;dr instead of taking implicit properties, we not take an implicit
Signal[Properties] (which will be provided by Rin's configuration code).

Also makes Signal/PeriodicSignal covariant, so that a Signal[A] is
also a Signal[B] if A is a subtype of B. This is needed for passing
implicit Signal[Properties] to the commons component.

Right now, to implement covariance I ended up using a lot of casting,
but I think this can be improved and I'll work on removing casting
in a future PR.