Add a controller to drive a distance with  a trapazoidal motion profile. Also add a task to use this controller, including a task with default values such as max acceleration and cruising velocity already set.
Refactor

Add a controller to drive a distance with  a trapazoidal motion profile. Also add a task to use this controller, including a task with default values such as max acceleration and cruising velocity already set.

Add a controller to drive a distance with  a trapazoidal motion profile. Also add a task to use this controller, including a task with default values such as max acceleration and cruising velocity already set.

Refactor