Current angle based direction reversal (#127)
* Refactor miscellaneous code, add lots of integration tests for pure
pursuit. Pure pursuit as of this stage is still broken.

* Make the idea of pure pursuit work for the case of driving backwards.
In case of failing unit tests, the robot wobbles as it approaches the
target, making the final angle not equal to the angle of the path
segment.

* Git done goofed and not added dis file

* Add even more integration tests. Failure to acheive the correct final
angle arises when traveling longer distances with pure pursuit than
rather than with shorter distances.

* Make unit tests stop early if task finishes early. Remove old hack to
stop applying angle correction when error is withing half the look
ahead distance.

* Tied left and right velocities to angular velocity in
simulator.

Integration tests pass because wobbling now stabilizes with the correct
physics. Despite this, wobble does exist initially even with
ridiculously large look ahead distances.

* wip

* Remove extra logging.

* Use absolute value in clamp

* Clean up imports and unnecessary explicit types

* Clean up tests

* Refactor simulator